Move from axum to tonic
This is a test prototype. Doesn't actually do anything related to raft yet

Rather than create REST endpoints using axum decided
to create RPC service with tonic. It will allow for easier
definitions of actions and models to be passed between the
raft nodes which I think will be helpful. Notes in obisdian on
more notes for how to accomplish this.