Tweak project structure
Plan now is to have the open to world client and the raft stuff both live in
raft-grpc rather than create a new package for the open to world client.
The risdb package will be what is vended (if at all) on crates. So, a client
that can be used by another service will be created in there