First REAL Raft Impl
Now leader election occurs and also data is replicated. This relies on a few things
though, first it relies on the client only ever sending writes to the actual leader.
It also assumes no nodes are brought back online. It can handle a node dying but
can't handle it coming back. I also haven't tested it all that extensively.
Regardless this is a pretty cool start and I'm feeling good about it