Small fixes
Mainly make it so that a leader converts to follower if a peer
responds with a higher term.

Overall this is in a pretty decent spot I'm super happy with progress
so far but it is not usable. Bringing servers back online only works
sometimes and there are certain scenarios where it just crashes and
I haven't fully worked out why. Also there is no log replay. So once
a new server comes online it doesn't catch up at all. Somehow it still
gets new values though? Don't understand that part