Swap from appending current logs directly
Previously was appending logs wrong where the current
set of logs was just getting sent no matter what. Instead
what should be happenign is the logs are written and
then logs are sent based on next_index. This swaps
to do that instead. Does not include any changes in
function other than that.