Make it so servers can be brought back online
These changes enable a server to crash and then be
fully caught back up again. Since logs and values
are stored in memory any time a server crashes it needs
to be fully brought back up to speed.

Still some bugs, one long todo for one of them was left
but this is a pretty big step in my opinion