Tests, minor startup change, and whitespace removal
Mainly added tests for the data-store. During these
tests changed the start-up to set a Noop log entry
when RaftImpl is initialized rather than in
the server start-up code. It made more sense that way
BUT it is still hacky. Have a todo to make it better.
Also made it so whitespace is trimmed on save so got
rid of some trailing whitespace.