Massive commit
Includes setup to make the Raft backend
something that can be interacted with via
just regular rust method calls and not
via localhost gRPC calls. This is good
setup for the next step of putting a hyper
server in front of the Raft stuff.

Also includes a lot of cleanup especially for
TODO comments