Another big commit
Make risdb-hyper into a library rather than
two binaries so that the code can be used
to front the Raft nodes. Kept what was
in the two binaries as examples in
examples/ folder. So far have the
handrolled client only implementing Get
with the Put to come next. Also right
now it is implemented since risdb-hyper
is just an echo service. Once it returns actual
responses I'll update the client

Also completely removes the old RisDbImpl that used
gRPC. :yipee: