Actually handle Get and Put
Make it so both Get and Put actually server
requests by interacting with the database.
Don't have a way to spin up a cluster
with hyper in front yet so haven't
dont a full test but will do that in the
next commit. Also may make the run()
method generic so it doesn't take RisDb
but instead an impl of ReadAndWrite to
make testing easier