Change how request id is set
Previously request id was set by clients
themselves which is dumb. Now it is set
automatically. This required refactoring
the inputs to the get and put methods
of RisDbClient since it no longer accepted
the ProtoBuf message type directly. Seems
better in the long run though as it makes
the customer facing method signature
more flexible