Working local version of risdb-deploy
With a loose definition of working. The logs are emitted and kept in a nice little
box which is cool. But the log lines don't wrap which makes them pretty useless
right now. I'll have to see what the best way to fix that is cause I don't think
letting lines wrap plays nice with superconsole but I'll have to look closer