It works! kinda
It works which is cool but it is super slow. It takes forever for a message to
get from the server stdout to the screen. So when I hit q to quit it can take like
10 seconds to actually quit. I feel like it should be instant and shouldn't add
much time to the overall process. I.e things should be drawn quickly after they are emitted
and processing a quick should also happen quickly